+++ services/ui-src/src/utils/index.ts
@@ -19,3 +19,4 @@ export * from "./other/focusElement";
 export * from "./other/scrollToTop";
 export * from "./other/time";
 export * from "./other/useBreakpoint";
+export * from "./other/useScrollPosition";
